requirements.txt
streamlit>=1.30.0
langchain>=0.1.0
langchain-community>=0.1.0
faiss-cpu>=1.7.4
google-generativeai>=0.4.0
python-dotenv>=1.0.0
gpt4all>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
